Update README.md
The objective of this project is therefore to develop a machine learning model capable of
classifying sounds according to the corresponding types of vehicles. The use of advanced signal
processing and classification techniques will create a robust and accurate system capable of
recognizing different types of vehicles, regardless of recording conditions.
By overcoming this problem, we will be able to benefit from many advantages, such as improved
traffic management, detection of unauthorized vehicles, monitoring of low emission zones, and
the implementation of effective strategies to reduce pollution. noise in urban areas.